striimu-ui-auth/backend/requirements.txt
Flask==2.0.1
PyJWT==2.1.0
